+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - `Innmind\TimeContinuum\Earth\FrozenClock` as been renamed `Innmind\TimeContinuum\Clock\Frozen`
 - `Innmind\TimeContinuum\Logger\Clock` as been renamed `Innmind\TimeContinuum\Clock\Logger`
 - Classes in the `Innmind\TimeContinuum\Earth\Move` namespace have been moved to `Innmind\TimeContinuum\Move`
+- `Innmind\TimeContinuum\Timezone` is now a `final class`
 
 ### Removed